+++ reco/src/main/java/com/example/reco/controllers/dto/RatingResponse.java
@@ -1,15 +1,16 @@
 package com.example.reco.controllers.dto;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public class RatingResponse {
     private final Long id;
     private final Long userId;
     private final Long itemId;
-    private final Short rating;
+    private final BigDecimal rating;
     private final Instant createdAt;
 
-    public RatingResponse(Long id, Long userId, Long itemId, Short rating, Instant createdAt) {
+    public RatingResponse(Long id, Long userId, Long itemId, BigDecimal rating, Instant createdAt) {
         this.id = id;
         this.userId = userId;
         this.itemId = itemId;
@@ -29,7 +30,7 @@ public class RatingResponse {
         return itemId;
     }
 
-    public Short getRating() {
+    public BigDecimal getRating() {
         return rating;
     }
